# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'ps-app-be'","description":"'Backend to support the PS web Site'","contact":{"name":"Hernán Jesús Morais","email":"104473@tecnicatura.frc.utn.edu.ar"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'Backend to support the PS web Site'"}],"paths":{"/ps/students/{id}":{"put":{"tags":["student-controller"],"operationId":"put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Student"}}}}}}},"/ps/users":{"get":{"tags":["user-controller"],"operationId":"getUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["user-controller"],"operationId":"postUserAddress","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/ps/students":{"get":{"tags":["student-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Student"}}}}}}},"post":{"tags":["student-controller"],"operationId":"post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Student"}}}}}}},"/ps/projects":{"get":{"tags":["project-controller"],"operationId":"getAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Project"}}}}}}},"post":{"tags":["project-controller"],"operationId":"postProject","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Project"}}}}}}},"/ps/invitations":{"get":{"tags":["invitation-controller"],"operationId":"getAll_2","parameters":[{"name":"user","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Invitation"}}}}}}},"post":{"tags":["invitation-controller"],"operationId":"sendInvitation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvitationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Invitation"}}}}}}},"/oauth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LogedUser"}}}}}}},"/ps/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/ps/invitations/{id}":{"get":{"tags":["invitation-controller"],"operationId":"getInvitation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Invitation"}}}}}}},"/ps/invitations/hash/{hash}":{"get":{"tags":["invitation-controller"],"operationId":"getInvitationWhitHash","parameters":[{"name":"hash","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Invitation"}}}}}}},"/config/code/frameworks":{"get":{"tags":["configuration-controller"],"operationId":"getAll_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CodeFramework"}}}}}}}}},"components":{"schemas":{"Address":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"street":{"type":"string"},"streetNumber":{"type":"string"},"zipCode":{"type":"string"},"detail":{"type":"string"},"city":{"type":"string"},"province":{"type":"string","enum":["B","K","H","U","C","W","X","E","P","Y","L","F","M","N","Q","R","A","J","D","Z","S","G","V","T"]},"country":{"type":"string","enum":["AR"]},"objectTypeName":{"type":"string"}}},"Contact":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"contactType":{"type":"string","enum":["EMAIL","PHONE","CELLPHONE"]},"value":{"type":"string"},"objectTypeName":{"type":"string"}}},"Grade":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"turn":{"type":"string"},"number":{"type":"string"},"name":{"type":"string"},"objectTypeName":{"type":"string"}}},"Identification":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"identification":{"type":"string"},"identificationType":{"type":"string","enum":["DNI","CUIL","CUIT","LEGAJO"]},"objectTypeName":{"type":"string"}}},"PlatformNetwork":{"type":"object","properties":{"internetPlatform":{"type":"string","enum":["LINKEDIN","FACEBOOK","INSTAGRAM","TWITTER","YOUTUBE","GITHUB","GITLAB","BITBUCKET","JIRA"]},"profileName":{"type":"string"}}},"Student":{"type":"object","allOf":[{"$ref":"#/components/schemas/Person"},{"type":"object","properties":{"universityIdentification":{"$ref":"#/components/schemas/Identification"},"careerAverage":{"type":"number"},"careerRegistrationDate":{"type":"string","format":"date"},"academicStatus":{"type":"string","enum":["REGULAR","LIBRE","INSCRIPTO"]},"grade":{"$ref":"#/components/schemas/Grade"}}}]},"Person":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"name":{"type":"string"},"lastName":{"type":"string"},"personIdentification":{"$ref":"#/components/schemas/Identification"},"address":{"$ref":"#/components/schemas/Address"},"birthday":{"type":"string","format":"date"},"age":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["CREATED","ACTIVE","INACTIVE"]},"personalContacts":{"type":"array","items":{"$ref":"#/components/schemas/Contact"}},"socialNetworks":{"type":"array","items":{"$ref":"#/components/schemas/PlatformNetwork"}},"universityContacts":{"type":"array","items":{"$ref":"#/components/schemas/Contact"}},"imageProfile":{"type":"string"},"objectTypeName":{"type":"string"}},"discriminator":{"propertyName":"objectType"}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"userName":{"type":"string"},"password":{"type":"string","writeOnly":true},"person":{"oneOf":[{"$ref":"#/components/schemas/Person"},{"$ref":"#/components/schemas/Student"}]},"roles":{"type":"array","items":{"type":"string","enum":["ADMIN","STUDENT","PROFESSOR"]}},"avatar":{"type":"string"},"objectTypeName":{"type":"string"}}},"Attachment":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"project":{"$ref":"#/components/schemas/Project"},"owner":{"oneOf":[{"$ref":"#/components/schemas/Person"},{"$ref":"#/components/schemas/Student"}]},"attachmentLink":{"type":"string"},"objectTypeName":{"type":"string"}}},"CodeFrameworkEntity":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"name":{"type":"string"},"type":{"type":"string"},"description":{"type":"string"},"imageLink":{"type":"string"},"iconLink":{"type":"string"}}},"CodeLanguageEntity":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"type":{"type":"string"},"name":{"type":"string"}}},"CodeRepository":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"internetPlatform":{"type":"string","enum":["LINKEDIN","FACEBOOK","INSTAGRAM","TWITTER","YOUTUBE","GITHUB","GITLAB","BITBUCKET","JIRA"]},"ownerName":{"type":"string"},"repositoryLink":{"type":"string"},"productionBranchName":{"type":"string"},"objectTypeName":{"type":"string"}}},"Comment":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"commentator":{"oneOf":[{"$ref":"#/components/schemas/Person"},{"$ref":"#/components/schemas/Student"}]},"comment":{"type":"string"},"objectTypeName":{"type":"string"}}},"Conversation":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"topic":{"type":"string"},"comments":{"type":"array","items":{"$ref":"#/components/schemas/Comment"}},"objectTypeName":{"type":"string"}}},"Issue":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"dueDate":{"type":"string","format":"date"},"issueStatus":{"type":"string","enum":["OPEN","PENDING","DONE"]},"summary":{"type":"string"},"resume":{"type":"string"},"informer":{"oneOf":[{"$ref":"#/components/schemas/Person"},{"$ref":"#/components/schemas/Student"}]},"responsible":{"oneOf":[{"$ref":"#/components/schemas/Person"},{"$ref":"#/components/schemas/Student"}]},"conversation":{"$ref":"#/components/schemas/Conversation"},"objectTypeName":{"type":"string"}}},"IssueTracker":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"internetPlatform":{"type":"string","enum":["LINKEDIN","FACEBOOK","INSTAGRAM","TWITTER","YOUTUBE","GITHUB","GITLAB","BITBUCKET","JIRA"]},"ownerName":{"type":"string"},"projectName":{"type":"string"},"boardLinks":{"type":"array","items":{"type":"string"}},"dashboardLinks":{"type":"array","items":{"type":"string"}},"objectTypeName":{"type":"string"}}},"Meeting":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"topic":{"type":"string"},"description":{"type":"string"},"meetingDateTime":{"type":"string","format":"date-time"},"requester":{"oneOf":[{"$ref":"#/components/schemas/Person"},{"$ref":"#/components/schemas/Student"}]},"guests":{"type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/Person"},{"$ref":"#/components/schemas/Student"}]}},"attendances":{"type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/Person"},{"$ref":"#/components/schemas/Student"}]}},"meetingLink":{"type":"string"},"wasMade":{"type":"boolean"},"minutesOfMeeting":{"$ref":"#/components/schemas/MinutesOfMeeting"},"objectTypeName":{"type":"string"}}},"MinutesOfMeeting":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"minutesOfMeeting":{"type":"string"},"objectTypeName":{"type":"string"}}},"PlatformEntity":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"type":{"type":"string"},"name":{"type":"string"}}},"Professor":{"type":"object","properties":{"universityIdentification":{"$ref":"#/components/schemas/Identification"},"universityContacts":{"type":"array","items":{"$ref":"#/components/schemas/Contact"}},"universityProfile":{"type":"string"}}},"Project":{"required":["description","limit","name","objective","projectType","scopes","students"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"name":{"maxLength":100,"minLength":10,"type":"string"},"description":{"maxLength":250,"minLength":10,"type":"string"},"objective":{"type":"string"},"limit":{"type":"string"},"scopes":{"type":"array","items":{"$ref":"#/components/schemas/ProjectScope"}},"projectType":{"type":"string","enum":["WEB","MOBILE","WEB_MOBILE"]},"projectStatus":{"type":"string","enum":["CREATED","UNDER_PROP_REVIEW","PROP_ACCEPTED","WIP","UNDER_FINAL_REVIEW","READY_TO_DELIVER","DELIVERED","FINISHED"]},"endDate":{"type":"string","format":"date"},"projectTheme":{"maxLength":100,"minLength":10,"type":"string"},"isRealProject":{"type":"boolean"},"imageLink":{"type":"string"},"issueTracker":{"$ref":"#/components/schemas/IssueTracker"},"codeRepositories":{"type":"array","items":{"$ref":"#/components/schemas/CodeRepository"}},"systemComponents":{"type":"array","items":{"$ref":"#/components/schemas/SystemComponent"}},"students":{"type":"array","items":{"$ref":"#/components/schemas/Student"}},"tutor":{"$ref":"#/components/schemas/Professor"},"observers":{"type":"array","items":{"$ref":"#/components/schemas/Professor"}},"valuations":{"type":"array","items":{"$ref":"#/components/schemas/Valuation"}},"academicCondition":{"type":"string","enum":["AC_1","AC_2"]},"meetings":{"type":"array","items":{"$ref":"#/components/schemas/Meeting"}},"conversation":{"$ref":"#/components/schemas/Conversation"},"issues":{"type":"array","items":{"$ref":"#/components/schemas/Issue"}},"attachments":{"type":"array","items":{"$ref":"#/components/schemas/Attachment"}},"projectPresentation":{"$ref":"#/components/schemas/ProjectPresentation"},"objectTypeName":{"type":"string"}}},"ProjectPresentation":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"presentationVideoLink":{"type":"string"},"demoVideoLink":{"type":"string"},"finalDocument":{"$ref":"#/components/schemas/Attachment"},"deliveryLink":{"type":"string"},"objectTypeName":{"type":"string"}}},"ProjectScope":{"required":["scope"],"type":"object","properties":{"project":{"$ref":"#/components/schemas/Project"},"scope":{"type":"string"}}},"SystemComponent":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"platformEntity":{"$ref":"#/components/schemas/PlatformEntity"},"technologyEntity":{"$ref":"#/components/schemas/TechnologyEntity"},"codeLanguageEntity":{"$ref":"#/components/schemas/CodeLanguageEntity"},"codeFrameworkEntity":{"$ref":"#/components/schemas/CodeFrameworkEntity"},"objectTypeName":{"type":"string"}}},"TechnologyEntity":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"type":{"type":"string"},"name":{"type":"string"}}},"Valuation":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"valuationType":{"type":"string","enum":["PRODUCT","MANAGEMENT","TOTAL_VALUATION"]},"value":{"type":"number"},"evaluator":{"$ref":"#/components/schemas/Professor"},"resume":{"type":"string"},"objectTypeName":{"type":"string"}}},"InvitationRequest":{"type":"object","properties":{"legajo":{"type":"string"},"email":{"type":"string"}}},"Invitation":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"legajo":{"type":"string"},"email":{"type":"string"},"hash":{"type":"string"},"link":{"type":"string"},"invitationStatus":{"type":"string","enum":["ACTIVE","USED","CANCELED","OVERDUE"]},"usedDate":{"type":"string","format":"date-time"},"dueDateTime":{"type":"string","format":"date-time"},"objectTypeName":{"type":"string"}}},"Login":{"type":"object","properties":{"userName":{"type":"string"},"password":{"type":"string"}}},"LogedUser":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"},"token":{"type":"string"}}},"CodeFramework":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"objectType":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"createdUser":{"type":"string"},"lastUpdatedDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deletedUser":{"type":"string"},"recordStatus":{"type":"string"},"name":{"type":"string"},"type":{"type":"string"},"description":{"type":"string"},"imageLink":{"type":"string"},"iconLink":{"type":"string"},"objectTypeName":{"type":"string"}}}}}}